+++ src/main/java/com/verdantartifice/thaumicwonders/common/entities/EntityHexamitePrimed.java
@@ -7,7 +7,6 @@ import net.minecraft.entity.MoverType;
 import net.minecraft.entity.item.EntityTNTPrimed;
 import net.minecraft.util.EnumParticleTypes;
 import net.minecraft.world.World;
-import net.minecraftforge.event.ForgeEventFactory;
 
 public class EntityHexamitePrimed extends EntityTNTPrimed {
     public EntityHexamitePrimed(World worldIn) {
@@ -53,10 +52,6 @@ public class EntityHexamitePrimed extends EntityTNTPrimed {
     }
     
     protected void explode() {
-        FluxExplosion explosion = new FluxExplosion(this.world, this, this.posX, this.posY + (double)(this.height / 16.0F), this.posZ, 6.0F, true, true, false);
-        if (!ForgeEventFactory.onExplosionStart(this.world, explosion)) {
-            explosion.doExplosionA();
-            explosion.doExplosionB(true);
-        }
+        FluxExplosion.create(this.world, this, this.posX, this.posY + (double)(this.height / 16.0F), this.posZ, 6.0F, true, true, false);
     }
 }