+++ src/main/java/com/verdantartifice/thaumicwonders/common/misc/FluxExplosion.java
@@ -2,17 +2,29 @@ package com.verdantartifice.thaumicwonders.common.misc;
 
 import java.util.List;
 
+import javax.annotation.Nullable;
+
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.Explosion;
 import net.minecraft.world.World;
+import net.minecraftforge.event.ForgeEventFactory;
 import thaumcraft.api.aura.AuraHelper;
 
 public class FluxExplosion extends Explosion {
     protected World world;
     protected boolean causesFlux;
+    
+    public static FluxExplosion create(World world, @Nullable Entity entity, double x, double y, double z, float strength, boolean isFlaming, boolean isSmoking, boolean causesFlux) {
+        FluxExplosion explosion = new FluxExplosion(world, entity, x, y, z, strength, isFlaming, isSmoking, causesFlux);
+        if (!ForgeEventFactory.onExplosionStart(world, explosion)) {
+            explosion.doExplosionA();
+            explosion.doExplosionB(true);
+        }
+        return explosion;
+    }
     
     public FluxExplosion(World worldIn, Entity entityIn, double x, double y, double z, float size, List<BlockPos> affectedPositions) {
         super(worldIn, entityIn, x, y, z, size, affectedPositions);